+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 0.0.1
+## 0.0.7
 
-* Initial setup of a background container
+* Container properties added to set padding, margin and decoration
 
 ## 0.0.3
 
 * Shapes in pattern made configurable
+
+## 0.0.1
+
+* Initial setup of a background container
+